+++ src/typedal/__init__.py
@@ -2,16 +2,15 @@
 TypeDAL Library.
 """
 
-from . import fields
-from .core import (
-    Relationship,
-    TypeDAL,
-    TypedField,
-    TypedRows,
-    TypedTable,
-    relationship,
-)
+from .core import TypeDAL
+from .fields import TypedField
 from .helpers import sql_expression
+from .query_builder import QueryBuilder
+from .relationships import Relationship, relationship
+from .rows import TypedRows
+from .tables import TypedTable
+
+from . import fields  # isort: skip
 
 try:
     from .for_py4web import DAL as P4W_DAL
@@ -19,6 +18,7 @@ except ImportError:  # pragma: no cover
     P4W_DAL = None  # type: ignore
 
 __all__ = [
+    "QueryBuilder",
     "Relationship",
     "TypeDAL",
     "TypedField",