+++ src/typedal/types.py
@@ -2,10 +2,18 @@
 Stuff to make mypy happy.
 """
 
-import typing
-from datetime import datetime
-from typing import Any, Callable, Optional, TypedDict
+# ---------------------------------------------------------------------------
+# Imports
+# ---------------------------------------------------------------------------
 
+# Standard library
+import datetime as dt
+import types
+import typing as t
+
+import pydal.objects
+
+# Third-party
 from pydal.adapters.base import BaseAdapter
 from pydal.helpers.classes import OpRow as _OpRow
 from pydal.helpers.classes import Reference as _Reference
@@ -13,132 +21,149 @@ from pydal.helpers.classes import SQLCustomType
 from pydal.objects import Expression as _Expression
 from pydal.objects import Field as _Field
 from pydal.objects import Query as _Query
+from pydal.objects import Row as _Row
 from pydal.objects import Rows as _Rows
 from pydal.objects import Set as _Set
 from pydal.objects import Table as _Table
 from pydal.validators import Validator as _Validator
-from typing_extensions import NotRequired
 
 try:
     from string.templatelib import Template
 except ImportError:
-    Template: typing.TypeAlias = str
+    Template: t.TypeAlias = str
 
-if typing.TYPE_CHECKING:
-    from .core import TypedField
+# Internal references
+if t.TYPE_CHECKING:
+    from .fields import TypedField
+    from .tables import TypedTable
 
 
-AnyDict: typing.TypeAlias = dict[str, Any]
+# ---------------------------------------------------------------------------
+# Aliases
+# ---------------------------------------------------------------------------
+
+AnyCallable: t.TypeAlias = t.Callable[..., t.Any]
+AnyDict: t.TypeAlias = dict[str, t.Any]
+
+
+# ---------------------------------------------------------------------------
+# Protocols
+# ---------------------------------------------------------------------------
+
+
+class TableProtocol(t.Protocol):  # pragma: no cover
+    """Protocol to make mypy happy for Tables."""
+
+    id: "TypedField[int]"
+
+    def __getitem__(self, item: str) -> "Field": ...
+
+
+class CacheFn(t.Protocol):
+    """
+    The cache model (e.g. cache.ram) accepts these parameters (all filled by default).
+    """
+
+    def __call__(
+        self: BaseAdapter,
+        sql: str = "",
+        fields: t.Iterable[str] = (),
+        attributes: t.Iterable[str] = (),
+        colnames: t.Iterable[str] = (),
+    ) -> "Rows": ...
+
+
+class FileSystemLike(t.Protocol):  # pragma: no cover
+    """Protocol for any class that has an 'open' function (e.g. OSFS)."""
+
+    def open(self, file: str, mode: str = "r") -> t.IO[t.Any]: ...
+
+
+# ---------------------------------------------------------------------------
+# pydal Wrappers (to help mypy understand these classes)
+# ---------------------------------------------------------------------------
 
 
 class Query(_Query):  # type: ignore
-    """
-    Pydal Query object.
-
-    Makes mypy happy.
-    """
+    """Pydal Query object. Makes mypy happy."""
 
 
 class Expression(_Expression):  # type: ignore
-    """
-    Pydal Expression object.
-
-    Make mypy happy.
-    """
+    """Pydal Expression object. Make mypy happy."""
 
 
 class Set(_Set):  # type: ignore
-    """
-    Pydal Set object.
-
-    Make mypy happy.
-    """
+    """Pydal Set object. Make mypy happy."""
 
 
-if typing.TYPE_CHECKING:
+if t.TYPE_CHECKING:
 
     class OpRow:
         """
         Pydal OpRow object for typing (otherwise mypy thinks it's Any).
-
-        Make mypy happy.
         """
 
-        def __getitem__(self, item: str) -> typing.Any:
-            """
-            Dict [] get notation.
-            """
+        def __getitem__(self, item: str) -> t.Any: ...
+        def __setitem__(self, key: str, value: t.Any) -> None: ...
 
-        def __setitem__(self, key: str, value: typing.Any) -> None:
-            """
-            Dict [] set notation.
-            """
-
-        # ... and more methods
+        # more methods could be added
 
 else:
 
     class OpRow(_OpRow):  # type: ignore
-        """
-        Pydal OpRow object at runtime just uses pydal's version.
-
-        Make mypy happy.
-        """
+        """Runtime OpRow, using pydal's version."""
 
 
 class Reference(_Reference):  # type: ignore
-    """
-    Pydal Reference object.
-
-    Make mypy happy.
-    """
+    """Pydal Reference object. Make mypy happy."""
 
 
 class Field(_Field):  # type: ignore
-    """
-    Pydal Field object.
-
-    Make mypy happy.
-    """
+    """Pydal Field object. Make mypy happy."""
 
 
 class Rows(_Rows):  # type: ignore
-    """
-    Pydal Rows object.
+    """Pydal Rows object. Make mypy happy."""
 
-    Make mypy happy.
-    """
-
-    def column(self, column: typing.Any = None) -> list[typing.Any]:
+    def column(self, column: t.Any = None) -> list[t.Any]:
         """
         Get a list of all values in a specific column.
 
         Example:
-                rows.column('name') -> ['Name 1', 'Name 2', ...]
+            rows.column('name') -> ['Name 1', 'Name 2', ...]
         """
         return [r[str(column) if column else self.colnames[0]] for r in self]
 
 
-class Validator(_Validator):  # type: ignore
-    """
-    Pydal Validator object.
+class Row(_Row): ...
 
-    Make mypy happy.
-    """
+
+class Validator(_Validator):  # type: ignore
+    """Pydal Validator object. Make mypy happy."""
+
+
+class Table(_Table, TableProtocol):  # type: ignore
+    """Table with protocol support. Make mypy happy."""
+
+
+# ---------------------------------------------------------------------------
+# Utility Types
+# ---------------------------------------------------------------------------
 
 
 class _Types:
-    """
-    Internal type storage for stuff that mypy otherwise won't understand.
-    """
+    """Internal type storage for stuff mypy otherwise won't understand."""
 
     NONETYPE = type(None)
 
 
-class Pagination(TypedDict):
-    """
-    Pagination key of a paginate dict has these items.
-    """
+# ---------------------------------------------------------------------------
+# TypedDicts
+# ---------------------------------------------------------------------------
+
+
+class Pagination(t.TypedDict):
+    """Pagination key of a paginate dict has these items."""
 
     total_items: int
     current_page: int
@@ -146,36 +171,30 @@ class Pagination(TypedDict):
     total_pages: int
     has_next_page: bool
     has_prev_page: bool
-    next_page: Optional[int]
-    prev_page: Optional[int]
+    next_page: t.Optional[int]
+    prev_page: t.Optional[int]
 
 
-class PaginateDict(TypedDict):
-    """
-    Result of PaginatedRows.as_dict().
-    """
+class PaginateDict(t.TypedDict):
+    """Result of PaginatedRows.as_dict()."""
 
     data: dict[int, AnyDict]
     pagination: Pagination
 
 
-class CacheMetadata(TypedDict):
-    """
-    Used by query builder metadata in the 'cache' key.
-    """
+class CacheMetadata(t.TypedDict):
+    """Used by query builder metadata in the 'cache' key."""
 
     enabled: bool
-    depends_on: list[Any]
-    key: NotRequired[str | None]
-    status: NotRequired[str | None]
-    expires_at: NotRequired[datetime | None]
-    cached_at: NotRequired[datetime | None]
+    depends_on: list[t.Any]
+    key: t.NotRequired[str | None]
+    status: t.NotRequired[str | None]
+    expires_at: t.NotRequired[dt.datetime | None]
+    cached_at: t.NotRequired[dt.datetime | None]
 
 
-class PaginationMetadata(TypedDict):
-    """
-    Used by query builder metadata in the 'pagination' key.
-    """
+class PaginationMetadata(t.TypedDict):
+    """Used by query builder metadata in the 'pagination' key."""
 
     limit: int
     current_page: int
@@ -184,110 +203,44 @@ class PaginationMetadata(TypedDict):
     min_max: tuple[int, int]
 
 
-class TableProtocol(typing.Protocol):  # pragma: no cover
-    """
-    Make mypy happy.
-    """
+class SelectKwargs(t.TypedDict, total=False):
+    """Possible keyword arguments for .select()."""
 
-    id: "TypedField[int]"
-
-    def __getitem__(self, item: str) -> Field:
-        """
-        Tell mypy a Table supports dictionary notation for columns.
-        """
-
-
-class Table(_Table, TableProtocol):  # type: ignore
-    """
-    Make mypy happy.
-    """
-
-
-class CacheFn(typing.Protocol):
-    """
-    The cache model (e.g. cache.ram) accepts these parameters (all filled by dfeault).
-    """
-
-    def __call__(
-        self: BaseAdapter,
-        sql: str = "",
-        fields: typing.Iterable[str] = (),
-        attributes: typing.Iterable[str] = (),
-        colnames: typing.Iterable[str] = (),
-    ) -> Rows:
-        """
-        Only used for type-hinting.
-        """
-
-
-# CacheFn = typing.Callable[[], Rows]
-CacheModel = typing.Callable[[str, CacheFn, int], Rows]
-CacheTuple = tuple[CacheModel, int]
-
-
-class SelectKwargs(TypedDict, total=False):
-    """
-    Possible keyword arguments for .select().
-    """
-
-    join: Optional[list[Expression]]
-    left: Optional[list[Expression]]
-    orderby: Optional[Expression | str | Table]
-    limitby: Optional[tuple[int, int]]
+    join: t.Optional[list[Expression]]
+    left: t.Optional[list[Expression]]
+    orderby: "OrderBy | t.Iterable[OrderBy] | None"
+    limitby: t.Optional[tuple[int, int]]
     distinct: bool | Field | Expression
     orderby_on_limitby: bool
     cacheable: bool
-    cache: CacheTuple
+    cache: "CacheTuple"
 
 
-class Metadata(TypedDict):
-    """
-    Loosely structured metadata used by Query Builder.
-    """
+class Metadata(t.TypedDict):
+    """Loosely structured metadata used by Query Builder."""
 
-    cache: NotRequired[CacheMetadata]
-    pagination: NotRequired[PaginationMetadata]
-
-    query: NotRequired[Query | str | None]
-    ids: NotRequired[str]
-
-    final_query: NotRequired[Query | str | None]
-    final_args: NotRequired[list[Any]]
-    final_kwargs: NotRequired[SelectKwargs]
-    relationships: NotRequired[set[str]]
-
-    sql: NotRequired[str]
+    cache: t.NotRequired[CacheMetadata]
+    pagination: t.NotRequired[PaginationMetadata]
+    query: t.NotRequired[Query | str | None]
+    ids: t.NotRequired[str]
+    final_query: t.NotRequired[Query | str | None]
+    final_args: t.NotRequired[list[t.Any]]
+    final_kwargs: t.NotRequired[SelectKwargs]
+    relationships: t.NotRequired[set[str]]
+    sql: t.NotRequired[str]
 
 
-class FileSystemLike(typing.Protocol):  # pragma: no cover
-    """
-    Protocol for any class that has an 'open' function.
-
-    An example of this is OSFS from PyFilesystem2.
-    """
-
-    def open(self, file: str, mode: str = "r") -> typing.IO[typing.Any]:
-        """
-        Opens a file for reading, writing or other modes.
-        """
-        ...
-
-
-AnyCallable: typing.TypeAlias = Callable[..., Any]
-
-
-class FieldSettings(TypedDict, total=False):
+class FieldSettings(t.TypedDict, total=False):
     """
     The supported keyword arguments for `pydal.Field()`.
-
     Other arguments can be passed.
     """
 
     type: str | type | SQLCustomType
     length: int
-    default: Any
+    default: t.Any
     required: bool
-    requires: list[AnyCallable | Any | Validator] | Validator | AnyCallable
+    requires: list[AnyCallable | t.Any | Validator] | Validator | AnyCallable
     ondelete: str
     onupdate: str
     notnull: bool
@@ -301,8 +254,8 @@ class FieldSettings(TypedDict, total=False):
     searchable: bool
     listable: bool
     regex: str
-    options: list[Any] | AnyCallable
-    update: Any
+    options: list[t.Any] | AnyCallable
+    update: t.Any
     authorize: AnyCallable
     autodelete: bool
     represent: AnyCallable
@@ -316,6 +269,46 @@ class FieldSettings(TypedDict, total=False):
     custom_delete: AnyCallable
     filter_in: AnyCallable
     filter_out: AnyCallable
-    custom_qualifier: Any
-    map_none: Any
+    custom_qualifier: t.Any
+    map_none: t.Any
     rname: str
+
+
+# ---------------------------------------------------------------------------
+# Generics & Query Helpers
+# ---------------------------------------------------------------------------
+
+T = t.TypeVar("T", bound=t.Any)
+P = t.ParamSpec("P")
+R = t.TypeVar("R")
+
+T_MetaInstance = t.TypeVar("T_MetaInstance", bound="TypedTable")
+T_Query = t.Union[
+    "Table",
+    Query,
+    bool,
+    None,
+    "TypedTable",
+    t.Type["TypedTable"],
+    Expression,
+]
+
+T_subclass = t.TypeVar("T_subclass", "TypedTable", Table)
+T_Field: t.TypeAlias = t.Union["TypedField[t.Any]", "Table", t.Type["TypedTable"]]
+
+# use typing.cast(type, ...) to make mypy happy with unions
+T_Value = t.TypeVar("T_Value")  # actual type of the Field (via Generic)
+
+# table-ish parameter:
+P_Table = t.Union[t.Type["TypedTable"], pydal.objects.Table]
+
+Condition: t.TypeAlias = t.Optional[t.Callable[[P_Table, P_Table], Query | bool]]
+
+OnQuery: t.TypeAlias = t.Optional[t.Callable[[P_Table, P_Table], list[Expression]]]
+
+CacheModel = t.Callable[[str, CacheFn, int], Rows]
+CacheTuple = tuple[CacheModel, int]
+
+OrderBy: t.TypeAlias = str | Expression
+
+T_annotation = t.Type[t.Any] | types.UnionType