+++ tests/test_main.py
@@ -1,16 +1,15 @@
 import re
+import typing
 import sys
 from copy import copy
 from sqlite3 import IntegrityError
 from typing import ForwardRef
 
-import pydal
 import pytest
 
-from src.typedal import *
+from src.typedal import TypedRows
 from src.typedal.__about__ import __version__
 from src.typedal.fields import *
-from typedal.types import Expression
 
 
 def test_about():